SAST/enhanced-cxflow-scaresolver/dispatcher/requirements.txt
pyyaml
python-daemon
psutil
docker
urllib3>=1.26.0
idna<4,>=2.5